requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
motor>=3.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
jinja2>=3.1.2
aiofiles>=23.1.0
pydantic>=2.0.0
bcrypt>=4.0.1 